+++ libs/core/src/lib/action-sheet/action-sheet-mobile/action-sheet-mobile.component.ts
@@ -1,23 +1,33 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, TemplateRef } from '@angular/core';
 
+type OpenChangeHandle = (isOpen: boolean) => void;
+
 @Component({
     selector: 'fd-action-sheet-mobile',
     templateUrl: './action-sheet-mobile.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ActionSheetMobileComponent {
-    /** Whenever links should be visible **/
+    /** Whenever links should be visible */
     open = false;
 
     /** @hidden */
     childContent: {
         actionSheetBodyTemplate: TemplateRef<any>;
+        isOpenChangeHandle: OpenChangeHandle;
     } = null;
 
-    constructor(private _changeDetectionRef: ChangeDetectorRef) {}
+    /** @hidden */
+    get _isOpenChangeHandle(): OpenChangeHandle {
+        return this.childContent.isOpenChangeHandle;
+    }
 
+    /** @hidden */
+    constructor(private readonly _changeDetectionRef: ChangeDetectorRef) {}
+
+    /** @hidden */
     toggleOpenState(isOpen: boolean): void {
         this.open = isOpen;
-        this._changeDetectionRef.detectChanges();
+        this._changeDetectionRef.markForCheck();
     }
 }