+++ libs/core/src/lib/notification/notification-body/notification-body.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectionStrategy, Component, HostBinding, ViewEncapsulation } from '@angular/core';
+import { NotificationGroupBaseDirective } from './../notification-utils/notification-group-base';
 
 @Component({
     selector: 'fd-notification-body',
@@ -6,7 +7,7 @@ import { ChangeDetectionStrategy, Component, HostBinding, ViewEncapsulation } fr
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class NotificationBodyComponent {
+export class NotificationBodyComponent extends NotificationGroupBaseDirective {
     /** @hidden */
     @HostBinding('class.fd-notification__body')
     fdNotificationBodyClass = true;