+++ libs/platform/src/lib/form/multi-input/base-multi-input.ts
@@ -158,12 +158,12 @@ export abstract class BaseMultiInput extends CollectionBaseInput implements Afte
     isOpenChange: EventEmitter<boolean> = new EventEmitter<boolean>();
 
     /** Event emitted when data loading is started. */
-    @Output()
-    onDataRequested = new EventEmitter<void>();
+    // eslint-disable-next-line @angular-eslint/no-output-on-prefix
+    @Output() onDataRequested = new EventEmitter<void>();
 
     /** Event emitted when data loading is finished. */
-    @Output()
-    onDataReceived = new EventEmitter<void>();
+    // eslint-disable-next-line @angular-eslint/no-output-on-prefix
+    @Output() onDataReceived = new EventEmitter<void>();
 
     /** @hidden */
     @ViewChild(ListComponent)